requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
PyJWT==2.9.0
SQLAlchemy==2.0.32
Werkzeug==3.0.3
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
itsdangerous==2.2.0
msgspec==0.18.6
pip==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8
snowflake-id==1.0.2
typing_extensions==4.12.2